+++ apps/pages/urls.py
@@ -1,5 +1,4 @@
 from django.urls import path
-
 from . import views
 
 
@@ -7,4 +6,6 @@ app_name = "pages"
 
 urlpatterns = [
     path('', views.index, name='index'),
+    path('healthz/', views.healthz, name='healthz'),
+    path('readiness/', views.readiness, name='readiness'),
 ]